[Refs #9,#10] Redo gitconfig as git-setup
Due to the versioning of gitconfig files, the gitconfig script for
installs was no longer necessary and works better better as a standalone
setup script.

Further, it's task is now different: rather than write the appropriate
values in the config file, it now simply confirms user/email address
settings (taking in new values as needed) and re-links the core.excludes
file, since the path `/Users/Knoble/` won't be the correct one for all
users.