[Refs #9,#10] Make install-all independent of loc
With these edits, install-all script is independent of where it is
placed and it also confirms the execution of scripts. Slightly less
attack prone.

This is necessary so that I can move it to a setup directory

Further, it fixes an issue where the directory itself was found by find
and couldn't be executed; now, find only retrieves files.