[Refs #9,#10] Edit README.md to reflect changes
Massive updates, including a tree output of the file structure, better
titles and organization, and the new usage information.

This was necessary due to the changes made in #10.

Some of the links are currently broken (those to the /docs folder)
because it does not exist yet. Covered as TODO in #12.

Close #5.