[Refs #13,#18] Only check folds entering a window
The cursor hold keeps it more automatic, but something in the script has
the poptential to "fail", in the sense that it causes the terminal to
complain/flash/alert me, which I don't like. I'd rather the script run
without said problem, and until I figure out how to fix it I'm just
turning off the CursorHold.