Make login Terminal command
When starting a Terminal session with the provided profile, the command
run by login will be login; that is, you will need to login to gain
access to your shell.

While *slightly* inconvenient, especially in that I can't run commands
from Alfred, I probably trust this a bit more and am happy to use it. Of
course, if someone wanted to avoid it, they could simply use another
Terminal profile, so it's not a security measure as much as a personal
feature.