Format install scripts
This is a combination of adding some extra echos and some extra newlines
within the scripts to help better visualize things.

It also helped to show me how hardcoded the scripts are, and yet I'm
struggling with allowing a different name for the directory. Modifying
all the right places during installation would be (a) tedious, (b) prone
to error, and (c) require everybody's git repo to be diverged from this
one, causing issues when updating. Instead, there could concievably be a
hidden file .dir that contained the name of the directory to use, and it
could be ignored somehow, but *that* is difficult due to the
configuration necessary. Plus it involves too many file reads. Still
looking for a solution.