Make HasFolds a little more modular
Now it has parameters that control what to set foldcolumn to, and
returns the value for you to use rather than setting foldcolumn itself.

This could be extended into a plugin Auto Fold Column, with autoloaded functions
that detect folds, return the right values (based on global variables,
options passed in, and sane defaults), and easy to set up with a simple

:call AutoFoldColumn() " Use defaults or global vars
:call AutoFoldColumn(col, def) " Use the variables passed in

AutoFoldColumn would set up the auto command.

Need to google if Vim has a concept of default parameters.