[vim] Do not lint on entering
Scriptease provides a :V family of commands that edit files on
'runtimepath' matching its arguments. It loads all matches into the
location list, which is overwritten on enter by ALE.

Turning off the on-enter functionality of ALE fixes this problem (though
not permanently--if I edit a file, ALE will trigger).

Other solutions would be
  - to only lint manually (blech)
  - to lint into the quickfix list (overrides make, vimgrep, :Ack, &c.)