dotfiles.plink: correct pre-reqs for gitk
$? was expanding to Makefile and Dracula/git/gitk.

By disabling the [] plink syntax, ~/.config/git/gitk depends only on
Dracula/git/gitk, allowing use of $? again.

In otherwords, if a plink/make recipe was relying on $? to only be one
file, well, that won't be true with the [] syntax (it will now include
the Makefile).

This may be a bug, or may not. Time will tell. For now, the solution is
to just let this little guy not depend on the Makefile.