vim: replace old "hacks" with my new plugin, simpl
All programmers should know the following about duplicate code:

If you write it once, leave it be.
If you write it twice, give it a second thought, but let it be.
If you write it a third time, refactor it.

Today, this happened.

I already had use and load set up for SML and lisp, respectively.

A class of mine will be using clojure shortly, and I knew I wanted to
extend this functionality to clojure, and maybe to other
REPL/interpreter-based languages eventually.

So it was time for the x3 refactor.

Unfortunately, the code was sort of scattered through different
filetypes, and keyed in to the old `terminal#run` function. So it took
some time to pull the pieces of that together and design the necessary
mechanisms by which to automate and configure everything.

But, today, vim-simpl was born. And boy-oh-boy is it beautiful.