vim + vifm: leave netrw enabled
I needed to leave the scp/network editing capabilities of netrw in,
which meant I could move all the config back to after.

Plus, I needed to define the commands after netrw (with !), so this was
necessary anyway.

Of note, neither :Vifm nor :e .'t do what I want. But if I just use
Explore &co., it won't be an issue.