vim: clojure: fix cond indent
letfn is borked: the indent forces all subforms (including the subforms
of the body!) to use 2-space indent, so

(letfn [(a [] 1)]
  (->> (a)
       inc))

becomes

(letfn [(a [] 1)]
  (->> (a)
    inc))

which is wrong and weird. However, leaving letfn out of
*g:clojure_special_indent_words* is even worse, as then the fn forms
don't get the right indent.

Solution? Use (let [f (fn f [,,,] ,,,)] ,,,). It's ugly, but it works.