bash: write gj
I found myself doing the following a lot:

1. git grep foo
2. <Esc>k0wijump <CR> (i.e., git jump grep foo)

I needed this to be easier. gj sovles this problem.

After running git grep, git diff, or even git merge, invoke gj to in
turn invoke git jump with the appropriate arguments.

It has to be a function, unfortunately, because scripts and subshell
contexts prevent

  history -p '!!'

from working.