vim: attempt to highlight code in liquid files
g:liquid_highlight_types is supposed to be for this (I think), but when
I set it to the same as the g:markdown_fenced_languages above, it causes
problems.

Why? As far as I can tell, the filetype is first set to markdown
(sourcing the highlights, including vim's); then, the filetype is set to
liquid (clearing all syntax, then sourcing the highlights, including
vim's); *then* the liquid subtype is set to markdown and used to source
more syntax (sourcing the highlights, including vim's a *second* time).

This second vim-syntax is what causes the problem, and I have no idea
how to fix it. Ironically, it actually works; that is, the code
highlighting is applied. It also throws an error at edit-time, which I
dislike. It appears to only happen with vim because of the line

  syn sync linecont	"^\s\+\\"

which is not allowed more than once (?).

The unfortunate temporary fix is to not have code-highlighting in liquid
files for the moment.