vim: set keywordprg to :Man in zsh files
The default :RunHelp tries to pipe the output into less, which I find to
be an abuse:

   command! -buffer -nargs=1 RunHelp silent exe '!zsh -ic "autoload -Uz run-help; run-help <args> 2>/dev/null | LESS= less"' | redraw!

That's what $MANPAGER is *for*; if I wanted less, I would use it
(assuming run-help is aliased to man, which it is by default).

Since I haven't setup run-help to do anything fancy yet (such as work
differently on builtins or functions or what-have-you), this will work
for the time being.

I might have to file a bug-report though to remove this less paging...