vim: prefer jobs over tmux when using :Make
vim-dispatch prefers tmux when available, but can asked to not use it
(which we do here). Jobs are faster and native.

Outside of tmux, this is a non-issue, but dispatch does the sensible
thing and checks for TMUX on it's own.

This affects :Dispatch and :Make commands. For :Start/:Spawn, using jobs
doesn't make sense, so they won't be used (and we'll get some crummy
fallback). Though, to be honest, I don't use them that often.