zsh: activate virtual env's earlier
In vim, python was always the system python (system() || !which ||
!python), even when in a virtualenv-activated directory and $PATH
contained $VIRTUAL_ENV/bin. (This screwed up &l:path computed by
apathy.)

The issue seemed to be difference in $PATH b/w vim and zsh -c (and it
was!). zsh -c is non-interactive, so ~/.zshrc doesn't run.

But ~/.zshrc is where we source .env/bin/activate (if it exists)!

So, we add that change to ~/.zshenv instead, so it's always found. And
we check first for $VIRTUAL_ENV, which is a more robust indicator of the
current virtual-env. We default to checking ./.env.

---

This took me ages to track down, since I track pretty closely path
changes (and can set path_debug=1 to show them on each shell
invocation).

But venv scripts don't use it! And source .env/bin/activate does no
reporting! So I didn't have good optics into this. Consider adding some
log info there (maybe a manual _path_debug call?).