vim: make dafny skolemization easier
It is common to grab a conjuct or disjunct involving quantifiers and
assert it, and then convert it to the corresponding statement form, when
one is writing a proof in dafny.

Two of the functions enable easy conversion between "assert quantifier"
form and "quantifier statement" form---the latter is what one typically
needs to carry out the proof. The statement form actually binds
variables, which is called skolemization.

A third function picks one of the two appropriately.

A last function, that is convenient in both the implementation of the
first two *and* when editing dafny foralls in general, toggles between
two forms of dafny foralls. In English, it's the difference between

  forall x, P(x) implies Q(x)

and

  forall x such that P(x) holds, Q(x) holds

Lastly, we take the obvious step of adding convenient mappings for these
functions (using popsikey! "q" is for "quantifier").