vim + markdown: add Navigate
Simplest heading navigator there ever was, with a twist: tries really
hard to get you on the heading you were on/under in the opened quickfix
window.

This turns out to be difficult; for the top of the file, we use the
first entry. For non-header lines, :cabove is a wonderful addition to
the toolkit. And for other lines, we search an approximation of the
entry in the quickfix and jump to it.