ale: use new ALELSPStarted autocommand
I've only mapped a few things so far; we'll see about others (I mapped a
lot of things when I set up OmniSharp). The gd and omnifunc stuff is
much nicer though.

I chose to leave Hover in the ALE popsikey: K is sometimes used for
this, but I like it better for real documentation, and Hover is a handy
way to "trigger" the LSP load if I feel like it's not coming online.

There's also other commands (CodeActions, etc.) that I haven't used much
and haven't mapped.