vim: hotload ftplugin/man.vim (:Man) when needed
This ftplugin is extremely slow to load the first time (~100ms), and I
may not even use it!

I recreate the command definition's attributes (see the inline script to
display them) and forward them to the actual command. The script does a
check for the command, so we delcommand it first. Subsequent invocations
don't need this.

With this, vim loads in ~100ms, down from ~300ms with man and tmplr (see
previous commit). That's ⅓ of the time, just by commenting out two lines
of code! Feels snappier already.