git hooks: handle git-worktrees correctly
`git rev-parse --git-dir` always reports the location of the `.git`
directory, but when using `git worktree` it is a file.
`--git-common-dir` reports the correct location.

Notably we could also use `git rev-parse --git-path hooks` or similar to
find specific files, but this does not work with `logs` (git always
reports the main worktree's logs).