vim: insert closers literally if no match
When typing one of the automatic closing brackets, occasionally it would
be preferable for the mapping to self-insert (*i.e.*, insert the key
pressed). This mostly comes up in comments, which are always considered
to not be a match, or when parentheses are unbalanced.

The `s:parenOfs` dict maintains a mapping of candidates to
"closeness" (?), which will be `-1` if the candidate should not be
considered. If all of the values are negative, then we should ignore the
closeness match and return the `map` argument, which is expected to be
the key(s) used in a mapping for self-insertion.