git: handle gron failure nicely
Some files advertise themselves as JSON and aren't (VsCode files are
notorious for this, apparently; they contain "comments" which are not in
the JSON spec).

This can cause failures when trying to run commands that trigger the
textconv (see `man gitattributes`).

The new JSON textconv silently falls back on reading the file normally
via `cat` if `gron` fails.