git: set commentChar = ;
This permits using # at the beginning of lines when convenient and is a
bit "lispier." I should not need ; at the beginning of a line for text,
but for Racket code I tend to use one ";".

Git 2.45.0 permits a multi-byte sequence, so I could use ";;" for
example. Better, it permits commentString as the variable name, so I can
set the latter to ";;" and maybe keep commentChar as ";" for old Git
(see git/git@9ccf3e9b22 (config: add core.commentString, 2024-03-27)).