vim: find wiki backlinks
This code uses some clever filename expansion: we expand to a full path
with :p, then substitute off whatever g:wiki_root is with :s?pat?repl?
and an empty replacement. This will break if your g:wiki_root has a ? or
pattern metacharacters, though :/ Finally, we lop off the leading slash
from absolute paths.

The search will also break if the filename has metacharactes in it for
:grep, which for me is rg at the moment.