zsh: make gj work with quoted patterns
When the patterns are quoted, the command expansion $(history -1 | …)
protects the resulting quotes so that git-jump sees them. Example

    λ git jump grep 'Ben Knoble'
    λ gj
    # nothing happens
    # ran the equivalent of
    λ git jump grep "'Ben Knoble'"

Since we're parsing history, `eval` the result (aka parse the quotes).
This seems like a slippery slope, but it works for now… let's find out
what other corner cases it creates.