git-div: avoid graphing base@{push} if it's the same as base
The same check can be added to base@{upstream}, but then we bail out
with a noisy message, which could be modified to say
    (maybe base, @{u}, and @{push} were all the same commit?)
but "silence is success" is nice. The main reason for this is so that
`git div` before a push shows something like I want from `git range-diff
@{push}...` but afterwards compares to @{upstream} so I can see what I'm
proposing. The old behavior (silence after a push) was less than useful
in the common case, leading me to write `git div @ @{u}` more than I'd
like.

Hopefully now `git div` "does what I mean."