vim: try harder to find long options
Git's manuals have a --[no-]long-opt convention that can make finding
long-opt with _ harder; add that to the pattern.

Also deal with Git's manuals that put the short and long forms on the
same line. See `git help grep` for some troublesome examples: we have
all of

       -A <num>, --after-context <num>
       -W, --function-context
       -O[<pager>], --open-files-in-pager[=<pager>]
       -<num>, -C <num>, --context <num>

Add a pattern for the short option that will work for all but the last
example.