git: write a script to split a topic into one branch per commit
The interactive rebase editor does most of the work, and getting that
right was the brunt of the pain.

So, though, was make `git rebase` get invoked with the correct
arguments! I want to support arbitrary <topic>s, but rebase takes
<upstream> <topic>, so I can't do `git rebase <maybe-upstream> <topic>`.
Ugh.

Instead, simulate the way git rebase checks out a topic with git
checkout. If rebase fails to find an upstream, this is annoying, but
`git switch -` gets you back. If rebase has conflicts, you're in the
expected state, and you don't want to automatically `git switch -`. What
a hack.

A future version might force you to pass upstream and topic through the
rebase options, and ask for a _description_ of a topic via --topic (only
for use with the default format string).