zsh: split most zshenv work to zshrc or zprofile
Some code (namely path manipulating functions) is duplicated because
functions are not inherited in the environment (nor, apparently, are -U
uniqueness flags on exported variables).
- The functions can be converted to autoload functions.
- The typeset declarations may need to persist.

However, all this duplication is because many of the "sourced" scripts
(but not all!) manipulate the environment! A better structure is likely
something like
- ~/.{,personal/}zsh/{,OS}/env/*zsh: sourced by a login shell; only used
  for effects which can be inherited.
- ~/.{,personal/}zsh/{,OS}/interactive/*zsh: sourced by interactive
  shells; used for effects which cannot be inherited (aliases,
  functions, etc.)
Let's try this reorganization, or something like it, before worrying
about centralizing pathadd &co. in autoload scripts: we may find that we
don't need to do that.