zsh: fix pathadd{,_front} for arguments with spaces
Apparently (q) doesn't quote spaces, but (qq) might. As a result,

    pathadd 'foo/ bar'

adds two elements (foo and bar) to path. Similarly for pathadd_front.

Rather than depend on re-quoting an argument, use eval _only_ to splice
in the variable name for which path is being manipulated. The rest is
expanded afterwards using the normal shell rules.