vim(wiki): don't try to commit commit-message files
When auto-committing, we edit a file that gets the filetype
wiki.gitcommit; that means that when writing it, we recursively try to
commit. This time, it fails horrifically (as opposed to the nice "not a
tracked file" type error). Besides, it seems error prone at best: avoid
committing git-like filetypes (though I may need to re-permit gitconfig
at some point).