git: add a faux-range-diff script based on a recent work experience
Say v1 had commits A and B. In the published v2, we made some changes,
then squashed, to form commit C. Now, without a high --creation-factor
we get something like

    git range-diff upstream v1 v2
    1: AAAA < -: ---- [subject]
    2: BBBB < -: ---- [subject]
    -: ---- > 1: CCCC [subject]

So instead we'd like to compare the files of interest that have changed.
Of course, if C is also rebased to a newer upstream than when A branched
off (IOW C^ is a descendant of A^), then a naive "git diff B C" or "git
diff --merge-base B C" includes irrelevant changes.

So use a technique like the one in faux-range-diff here to limit the
diff to relevant files. If those files have changed irrelevantly in
other commits, this will show them (unhelpfully). In that case, consider
trying to make sense of a high-creation-factor range-diff.