zsh: complete tags from my tagsfile
My hooks usually write to .git/tags, so point Zsh at it (see
_complete_tag, bound to ^Xt, in zshall(1)).

Of course, if .git is not a directory or where the data is, this won't
work… if such a thing happens, I might have to set a hook on changing
directories to reset tagsfile to $(git rev-parse …)/tags or similar.
Fortunately Zsh does the upward search for me.